backend/python_image_service/requirements.txt
fastapi
uvicorn
pillow
requests
python-multipart
TTS
